Move color and tint into new file
