+++ package.json
@@ -51,6 +51,7 @@
     "src/animation.js",
     "src/sprite-animation.js",
     "src/drawing.js",
+    "src/color.js",
     "src/isometric.js",
     "src/diamondiso.js",
     "src/particles.js",
